espark-core/requirements.dev.txt
aiosqlite==0.22.1
autopep8==2.3.2
build==1.3.0
fastapi-cli[standard-no-fastapi-cloud-cli]==0.0.23
httpx==0.28.1
pycodestyle==2.14.0
pylint==4.0.4
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
twine==6.2.0
